Tests: rename the test class/file to `TokenizeTest`
Realistically, all tests in the `HighligherTest` file are testing the handling of various token types and only incidentally testing the rest of the logic of the `Highlighter` class.

With that in mind, I'm proposing to rename the test file to make it more obvious what is being tested. This will then also allow for adding additional new test classes to test the rest of the functionality of the `Highlighter` class.

Includes adding minimal documentation and a minor method order tweak.