TokenizeTest: add additional test cases for text string handling
This adds extra tests for:
* Double quoted strings with interpolated variables.
* Nowdoc
* Heredocs with and without interpolated variables.

Note: for the nowdoc and heredoc tests, the test input is put in single quotes to prevent "nested here/nowdoc", which could skew the tests.
These test cases also need to have a new line after the `;` as otherwise PHP does not tokenize these correctly (and the highlighter is not concerned with fixing incorrect tokenization of PHP itself).