Highlighter::getCodeSnippet(): bug fix - too many lines retrieved at start of file
When a code snippet in the middle of code is retrieved, the target snippet length calculation `$length = $linesAfter + $linesBefore + 1;` functions correctly, displaying the target line in the middle, padded by x number of lines before and after.
Given `$linesBefore` and `$linesAfter` both being set to two, this would display as:
```
before
before
target
after
after
```

Similarly, when a code snippet at the end of the code is retrieved, it will work correctly as well, the target line is displayed with x number of lines before. The number of lines _after_ is limited automatically by the length of the `$tokenLines` array.
Given `$linesBefore` and `$linesAfter` both being set to two, this would display as:
```
before
before
target
```

However, when a code snippet at the _start_ of the code is retrieved, the number of lines retrieved was incorrect. The target line would display with `($linesBefore + $linesAfter)` lines _after_ the target line.

Given `$linesBefore` and `$linesAfter` both being set to two, this currently displays as:
```
target
after
after
after
after
```

... while it should be:
```
target
after
after
```

The change in this commit fixes this bug.